# smart-home-microservices/internal/devices/adapters/pgrepo/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package pgrepo

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Device struct {
	ID        string
	Type      string
	Name      string
	Online    bool
	OnOff     bool
	UserID    string
	HomeID    pgtype.Text
	CreatedAt pgtype.Timestamp
}

type DevicesCreatedOutbox struct {
	DeviceID pgtype.Text
}

type DevicesDeletedOutbox struct {
	DeviceID pgtype.Text
}

type DevicesUpdatedOutbox struct {
	DeviceID pgtype.Text
}

type Home struct {
	ID        string
	Name      string
	UserID    string
	CreatedAt pgtype.Timestamp
}